requirements.txt
psycopg2==2.9.10
pytz==2024.1
python-dotenv==1.1.0
